pyproject.toml
[project]
name = "vectordb-learn"
version = "0.1.0"
description = "Educational tool for learning PostgreSQL vector databases"
requires-python = ">=3.13"
dependencies = [
    "textual>=1.0.0",
    "rich>=13.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-ai>=0.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "ollama>=0.3.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "structlog>=24.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true